requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
flask-migrate
flask-login
flask-bcrypt
python-dotenv